# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CariMurah.ai API",
    "description": "API for CariMurah.ai - Indonesian Shopping Agent with MongoDB storage",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://carimurah-ai-875241790017.europe-west1.run.app",
      "description": "Production Server (Cloud Run)"
    },
    {
      "url": "https://ais-pre-7xkvumwn2pakj7laysavzs-111006670209.asia-southeast1.run.app",
      "description": "Shared App Server (AIS Preview)"
    }
  ],
  "paths": {
    "/api/profile/{uid}": {
      "get": {
        "summary": "Retrieve user profile and preferences",
        "description": "Gets the user's stored profile from MongoDB, including their B2B/B2C focus (price, delivery, or rating) and subscription level.",
        "operationId": "getUserProfile",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "description": "Unique Firebase User ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "User profile data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserProfile" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create or update user profile",
        "description": "Saves user preferences to MongoDB. Use this to set whether the user prefers retail (B2C) or wholesale (B2B) findings.",
        "operationId": "updateUserProfile",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "description": "Unique Firebase User ID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "description": "Profile data to save",
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserProfile" }
            }
          }
        },
        "responses": {
          "200": { "description": "Profile updated successfully" }
        }
      }
    },
    "/api/history": {
      "get": {
        "summary": "Get global or sample shopping history",
        "description": "Returns a list of recent analysis results. Can be filtered by uid as a query parameter.",
        "operationId": "getGlobalHistory",
        "parameters": [
          {
            "name": "uid",
            "in": "query",
            "description": "Optional user ID to filter findings",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of findings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/HistoryItem" }
                }
              }
            }
          }
        }
      }
    },
    "/api/history/{uid}": {
      "get": {
        "summary": "Get specific user shopping history",
        "description": "Returns the list of all products found and analyzed for a specific user from MongoDB.",
        "operationId": "getUserHistory",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "description": "Firebase User ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "User history data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/HistoryItem" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Save a finding to user history",
        "description": "Persists an analysis result (product, price, platform, and savings) to the user's permanent history in MongoDB.",
        "operationId": "saveToHistory",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "description": "Firebase User ID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "description": "The finding details to save",
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/HistoryItemCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "id": { "type": "string" } }
                }
              }
            }
          }
        }
      }
    },
    "/api/history/{uid}/{historyId}": {
      "patch": {
        "summary": "Update specific history item",
        "description": "Update notes or details for a previously saved finding.",
        "operationId": "updateHistoryItem",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "historyId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "description": "Update successful" }
        }
      }
    },
    "/api/process": {
      "post": {
        "summary": "Process image or text for price analysis",
        "description": "Main agentic workflow. Analyze search queries or receipt images to find best prices on Tokopedia/Shopee.",
        "operationId": "processAnalysis",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": { "type": "string", "description": "Search query or description of the product" },
                  "image": { "type": "string", "description": "Optional Base64 encoded image (JPEG) of product or receipt" },
                  "isB2B": { "type": "boolean", "description": "Set true to look for Grosir/Wholesale prices" },
                  "preferences": { "$ref": "#/components/schemas/UserProfile" }
                },
                "required": ["text"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Detailed multi-agent analysis result" }
        }
      }
    },
    "/api/tts": {
      "post": {
        "summary": "Convert text to Zephyr AI voice",
        "description": "Generates high-quality studio audio for shopping summaries using the Zephyr persona voice.",
        "operationId": "textToSpeech",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": { "type": "string", "description": "Text to read aloud" },
                  "voice": { "type": "string", "default": "Zephyr", "enum": ["Zephyr", "Puck", "Charon"] }
                },
                "required": ["text"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Base64 encoded audio data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "audio": { "type": "string", "description": "Base64 string" },
                    "rate": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "summary": "Service status",
        "operationId": "checkHealth",
        "responses": {
          "200": { "description": "Active" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserProfile": {
        "type": "object",
        "properties": {
          "uid": { "type": "string" },
          "displayName": { "type": "string" },
          "email": { "type": "string" },
          "b2bFocus": { "type": "string", "enum": ["price", "delivery", "rating"] },
          "subscription": {
            "type": "object",
            "properties": {
              "tier": { "type": "string", "enum": ["FREE", "PRO", "ENTERPRISE"] }
            }
          }
        }
      },
      "HistoryItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "date": { "type": "string" },
          "itemsCount": { "type": "integer" },
          "totalSaved": { "type": "number" },
          "type": { "type": "string", "enum": ["B2C", "B2B"] },
          "productName": { "type": "string" },
          "brand": { "type": "string" },
          "recommendedPrice": { "type": "number" },
          "platform": { "type": "string" }
        }
      },
      "HistoryItemCreate": {
        "type": "object",
        "required": ["productName", "brand", "recommendedPrice", "platform"],
        "properties": {
          "productName": { "type": "string" },
          "brand": { "type": "string" },
          "recommendedPrice": { "type": "number" },
          "platform": { "type": "string" },
          "totalSaved": { "type": "number" },
          "type": { "type": "string", "default": "B2C" }
        }
      }
    }
  }
}